+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 # Unreleased
 
+- Auto derive the fee payer address when simulating a transaction with the fee payer public key
+
 # 1.33.1 (2024-11-28)
 
 - Add `gasProfile` function to `Move` class to allow for gas profiling of Aptos Move functions